fixed glitched mouse movement and added lights to hallway procgen (colors are glitching)

+++ lib/controls/controls.h
@@ -37,16 +37,16 @@ public:
 
         case SDL_EVENT_MOUSE_BUTTON_DOWN:
           if (event.button.button == SDL_BUTTON_LEFT)
-            mouse.down((int) event.button.x, (int) event.button.y);
+            mouse.down(event.button.x, event.button.y);
           break;
 
         case SDL_EVENT_MOUSE_MOTION:
-          mouse.move((int) event.motion.x, (int) event.motion.y, (int) event.motion.xrel, (int) event.motion.yrel);
+          mouse.move(event.motion.x, event.motion.y, event.motion.xrel, event.motion.yrel);
           break;
 
         case SDL_EVENT_MOUSE_BUTTON_UP:
           if (event.button.button == SDL_BUTTON_LEFT)
-            mouse.up((int) event.button.x, (int) event.button.y);
+            mouse.up(event.button.x, event.button.y);
           break;
 
         case SDL_EVENT_WINDOW_RESIZED:
